Add structs in module
