--- Source/Models/PHSettings.swift
@@ -1,15 +1,0 @@
-//
-//  PHSettings.swift
-//  Product Hunt
-//
-//  Created by Vlado on 4/27/16.
-//  Copyright © 2016 ProductHunt. All rights reserved.
-//
-
-import Foundation
-
-struct PHSettings {
-    var autologinEnabled: Bool
-    var showsCount: Bool
-    var filterCount: Int
-}
